RL/requirements.txt
absl-py==2.1.0
ale-py==0.8.1
annotated-types==0.6.0
anyio==4.3.0
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.1
Farama-Notifications==0.0.4
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
gensim==4.3.2
grpcio==1.63.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.8
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.12.0
Shimmy==1.3.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
stable_baselines3==2.3.2
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
transformers~=4.41.1
nltk~=3.8.1